requirements.txt
certifi==2023.11.17
chardet==3.0.4
fuzzywuzzy==0.18.0
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.1.5
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
iniconfig==2.0.0
Levenshtein==0.23.0
numpy==1.26.3
opencv-python==4.9.0.80
packaging==23.2
pillow==10.2.0
pluggy==1.3.0
pyenchant==3.2.2
pytesseract==0.3.10
pytest==7.4.4
python-Levenshtein==0.23.0
rapidfuzz==3.6.1
rfc3986==1.5.0
sniffio==1.3.0
